Camera addition + minimap
Player can now zoom with the scroll wheel. It's still rough though (zoom
out faster than zoom in for some reason) I will try to find out how to
best use SmoothDamp for this. There is also a minimap, only thing is the
lighting: it's receiving the main lighting, plus I added a spotlight but
that affects the rest of the game :/