Now clicking outside the button panel makes it disappear
I didn't come up with these codes; I found them on the Unity forums.
Long story short, they create an invisible background and set its Rect
Transform to be the same as the UI element (here, the button panel).
This background is then enabled whenever the buttons appear. On the
button panel, the MenuVisibilityCtrl script is attached: it sets the
button panel as the parent of the background, thus disabling it when we
make the buttons disappear. The main thing to remember is that the
background is assigned on Awake(), therefore the Popup Button panel
needs to be ENABLED in the inspector; it will disappear when we start
the scene!