Tower costs/refunds + other changes
Towers now have a "ladder" for costs and refunds. Creating more
increases cost, destroying gives it back. There are some major changes
in the GameController script, in the section that handles tower
creation: it now has a TowerChange function, hopefully its a good way to
go at it. (Check the prefabs to edit the cost lists!) (Also,
ScoreManager and TowerManager have been removed)