Tower range circles
A few things for this: I had trouble getting the scripts to communicate
on this. The code is "pretty" for disabling when clicking buttons,
however when clicking outside the button panel it gets messier. That's
because the invisble bg scripts affect only the UI panels, but the line
renderer is a component of the Tower, which is a child of the spot, and
only the GameController is aware of what spot is being selected. So...
yeah. I can change it if anyone comes up with a better way.